Address PR comments

+++ src/main/java/uk/gov/hmcts/appregister/applicationentry/dto/ApplicationListEntryDto.java
@@ -1,5 +1,6 @@
 package uk.gov.hmcts.appregister.applicationentry.dto;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import uk.gov.hmcts.appregister.applicationcode.dto.ApplicationCodeDto;
 import uk.gov.hmcts.appregister.common.enumeration.FeeStatusType;
@@ -21,6 +22,6 @@ public record ApplicationListEntryDto(
         String notes,
         String bulkUpload,
         Long resultId,
-        Long changedBy,
+        BigDecimal changedBy,
         OffsetDateTime changedDate,
         Long version) {}
